+++ src/constants/fetchStatus.ts
@@ -1,0 +1,8 @@
+enum FetchStatus {
+  UNFETCHED = 'UNFETCHED',
+  FETCHED = 'FETCHED',
+  FETCHING = 'FETCHING',
+  ERROR = 'ERROR',
+}
+
+export default FetchStatus;
